split fiels image and delete empty fiels and handel error quote> releats #18 #24 #25 #26
