+++ package.json
@@ -10,7 +10,7 @@
     "test": "NODE_ENV=test node tests/queries.test.js | tap-spec",
     "buildDB": "NODE_ENV=test node src/model/database/db_build.js"
   },
-  "pre-commit":[
+  "pre-commit": [
     "lint"
   ],
   "repository": {
