+++ src/views/error.hbs
@@ -1,0 +1,5 @@
+<section class="error">
+  <p class="error__desc">{{statusCode}}.
+    <span class="error__msg">{{errorMessage}}</span>
+  </p>
+</section>
