+++ src/views/home.hbs
@@ -9,7 +9,7 @@
 			<button class="btn-like">
 			</form>
 				<span id="count">{{this.likes}}</span>
-				<img src="images/like1.png" alt="like">
+				<img src="assets/icons/like1.png" alt="like">
 			</button>
 		</div>
 	</div>
