+++ src/views/layouts/error.hbs
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+  {{>htmlHead}}
+<body>
+
+  <main class="main">
+  {{{body}}}
+  </main>
+
+</body>
+</html>
