+++ src/views/partials/form-data.hbs
@@ -1,6 +1,6 @@
 	<form class="data-form" action="/addpic" method="POST" enctype="multipart/form-data">
 		<p id="para">add post</p><br>
-		<input class="txt" type="text" name="title" placeholder="Title" maxlength="20"><br><br>
+		<input class="txt" type="text" name="title" placeholder="Title" maxlength="20" required><br><br>
 		<input class="txt" type="text" name="tags" placeholder="Tags" maxlength="20"><br><br>
 		<div class="btn-post">
 			<div class="custom-file">
